Resolve API call issue

+++ app/src/main/java/com/dev/aman/movies_tmdb/data/model/NowPlaying.kt
@@ -48,7 +48,7 @@ data class NowPlaying(
         @SerializedName("video")
         val video: Boolean?,
         @SerializedName("vote_average")
-        val voteAverage: Int?,
+        val voteAverage: Double?,
         @SerializedName("vote_count")
         val voteCount: Int?
     )
